+++ agave_app/qtControls/RangeWidget.cpp
