+++ agave_app/qtControls/RangeWidget.h
