+++ agave_app/qtControls/Section.cpp
